New backups: RestoreBackUpUseCase + changes to MetadataHandler
This PR introduces a use case for restoring data from a backup to the database given a file previously created with CreateBackUpUseCase.
The steps from CreateBackUpUse case are repeated with the difference being that here we check if we are able to handle the backup version,
written down in metadata.json. In practice, since this is the first ever backup version, we only check if the version from the file is the same
as our current version, namely 1. In future, when we introduce a new version, we will also have to introduce a step for migrating older
backups.